dodanie słowników, zmiana Integer na Long, dodanie mapowania z konwencji camelCase na konwencję odpowiednią dla API

+++ src/main/java/pl/bestdinner/controller/DishController.java
@@ -30,19 +30,19 @@ public class DishController {
     }
 
     @DeleteMapping(value = "/{id}")
-    public ResponseEntity delete(@PathVariable("id") Integer id){
+    public ResponseEntity delete(@PathVariable("id") Long id){
         service.delete(id);
         return new ResponseEntity(HttpStatus.OK);
     }
 
     @GetMapping(value = "/{id}")
-    public ResponseEntity<DishDto> get(@PathVariable("id") Integer id){
+    public ResponseEntity<DishDto> get(@PathVariable("id") Long id){
         return new ResponseEntity<>(service.get(id), HttpStatus.OK);
     }
 
     @PutMapping(value = "/{id}")
     public ResponseEntity<DishDto> update(@RequestBody DishDto requestBody,
-                                            @PathVariable("id") Integer id){
+                                            @PathVariable("id") Long id){
         return new ResponseEntity<>(service.update(requestBody, id), HttpStatus.OK);
     }
 }
